Move kill process tree and re-use it
